Contactos_JCR-main/backend/requirements.txt
# Framework y servidor
fastapi==0.115.12
uvicorn==0.34.2

# Base de datos
sqlalchemy==2.0.40
# aiosqlite solo si usas async con SQLite
aiosqlite==0.19.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
# bcrypt solo si lo usas directamente, passlib[bcrypt] ya lo incluye
# bcrypt==4.0.1
python-multipart==0.0.20

# Validación y configuración
pydantic==2.11.3
pydantic-settings==2.1.0
email-validator==2.2.0
python-dotenv==1.0.0

# Testing
pytest==8.0.0
httpx==0.26.0
pytest-cov==4.1.0